+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.2.0] - 2020-08-26
+### Add validator for NZ accounts
+* Add the possibility to pass 'nz' as country option and validate the account number for this country.
+
 ## [1.1.0] - 2020-08-26
 ### Add options param
 * Add the possibility to pass an options object with the institution name. Validator will check the correctness of the ccc number and if